Spit attachment tests
