+++ rest/blip_api_test.go
@@ -13,12 +13,10 @@ package rest
 import (
 	"encoding/base64"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"log"
 	"net/http"
 	"net/url"
-	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -27,8 +25,6 @@ import (
 
 	"github.com/couchbase/go-blip"
 	"github.com/couchbase/gocb/v2"
-	"github.com/couchbase/gocbcore/v10/memd"
-	sgbucket "github.com/couchbase/sg-bucket"
 	"github.com/couchbase/sync_gateway/base"
 	"github.com/couchbase/sync_gateway/db"
 	"github.com/stretchr/testify/assert"
@@ -1234,12 +1230,6 @@ func AssertChangeEquals(t *testing.T, change []interface{}, expectedChange Expec
 	}
 }
 
-// Test adding / retrieving attachments
-func TestAttachments(t *testing.T) {
-	// TODO: Write tests to cover scenario
-	t.Skip("not tested")
-}
-
 // Make sure it's not possible to have two outstanding subChanges w/ continuous=true.
 // Expected behavior is that the first continous change subscription should get discarded in favor of 2nd.
 func TestConcurrentChangesSubscriptions(t *testing.T) {
@@ -1520,182 +1510,6 @@ func TestCheckpoint(t *testing.T) {
 	assert.True(t, strings.Contains(string(body), "Key"))
 	assert.True(t, strings.Contains(string(body), "Value"))
 
-}
-
-// Test Attachment replication behavior described here: https://github.com/couchbase/couchbase-lite-core/wiki/Replication-Protocol
-// - Put attachment via blip
-// - Verifies that getAttachment won't return attachment "out of context" of a rev request
-// - Get attachment via REST and verifies it returns the correct content
-func TestPutAttachmentViaBlipGetViaRest(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyHTTP, base.KeySync, base.KeySyncMsg)
-
-	// Create blip tester
-	bt, err := NewBlipTesterFromSpec(t, BlipTesterSpec{
-		connectingUsername: "user1",
-		connectingPassword: "1234",
-	})
-	require.NoError(t, err, "Unexpected error creating BlipTester")
-	defer bt.Close()
-
-	attachmentBody := "attach"
-	digest := db.Sha1DigestKey([]byte(attachmentBody))
-
-	input := SendRevWithAttachmentInput{
-		docId:            "doc",
-		revId:            "1-rev1",
-		attachmentName:   "myAttachment",
-		attachmentLength: len(attachmentBody),
-		attachmentBody:   attachmentBody,
-		attachmentDigest: digest,
-	}
-	bt.SendRevWithAttachment(input)
-
-	// Try to fetch the attachment directly via getAttachment, expected to fail w/ 403 error for security reasons
-	// since it's not in the context of responding to a "rev" request from the peer.
-	getAttachmentRequest := blip.NewRequest()
-	getAttachmentRequest.SetProfile(db.MessageGetAttachment)
-	getAttachmentRequest.Properties[db.GetAttachmentDigest] = input.attachmentDigest
-	getAttachmentRequest.Properties[db.GetAttachmentID] = input.docId
-	sent := bt.sender.Send(getAttachmentRequest)
-	if !sent {
-		panic(fmt.Sprintf("Failed to send request for doc: %v", input.docId))
-	}
-	getAttachmentResponse := getAttachmentRequest.Response()
-	errorCode, hasErrorCode := getAttachmentResponse.Properties["Error-Code"]
-	assert.Equal(t, "403", errorCode) // "Attachment's doc not being synced"
-	assert.True(t, hasErrorCode)
-
-	// Get the attachment via REST api and make sure it matches the attachment pushed earlier
-	response := bt.restTester.SendAdminRequest("GET", fmt.Sprintf("/db/%s/%s", input.docId, input.attachmentName), ``)
-	assert.Equal(t, input.attachmentBody, response.Body.String())
-
-}
-
-func TestPutAttachmentViaBlipGetViaBlip(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyHTTP, base.KeySync, base.KeySyncMsg)
-
-	// Create blip tester
-	bt, err := NewBlipTesterFromSpec(t, BlipTesterSpec{
-		connectingUsername:          "user1",
-		connectingPassword:          "1234",
-		connectingUserChannelGrants: []string{"*"}, // All channels
-	})
-	require.NoError(t, err, "Unexpected error creating BlipTester")
-	defer bt.Close()
-
-	attachmentBody := "attach"
-	digest := db.Sha1DigestKey([]byte(attachmentBody))
-
-	// Send revision with attachment
-	input := SendRevWithAttachmentInput{
-		docId:            "doc",
-		revId:            "1-rev1",
-		attachmentName:   "myAttachment",
-		attachmentLength: len(attachmentBody),
-		attachmentBody:   attachmentBody,
-		attachmentDigest: digest,
-	}
-	sent, _, _ := bt.SendRevWithAttachment(input)
-	assert.True(t, sent)
-
-	// Get all docs and attachment via subChanges request
-	allDocs, ok := bt.WaitForNumDocsViaChanges(1)
-	require.True(t, ok)
-
-	// make assertions on allDocs -- make sure attachment is present w/ expected body
-	require.Len(t, allDocs, 1)
-	retrievedDoc := allDocs[input.docId]
-
-	// doc assertions
-	assert.Equal(t, input.docId, retrievedDoc.ID())
-	assert.Equal(t, input.revId, retrievedDoc.RevID())
-
-	// attachment assertions
-	attachments, err := retrievedDoc.GetAttachments()
-	assert.True(t, err == nil)
-	assert.Equal(t, 1, len(attachments))
-	retrievedAttachment := attachments[input.attachmentName]
-	require.NotNil(t, retrievedAttachment)
-	assert.Equal(t, input.attachmentBody, string(retrievedAttachment.Data))
-	assert.Equal(t, len(attachmentBody), retrievedAttachment.Length)
-	assert.Equal(t, retrievedAttachment.Digest, input.attachmentDigest)
-
-}
-
-// Reproduces the issue seen in https://github.com/couchbase/couchbase-lite-core/issues/790
-// Makes sure that Sync Gateway rejects attachments sent to it that does not match the given digest and/or length
-func TestPutInvalidAttachment(t *testing.T) {
-
-	tests := []struct {
-		name                  string
-		correctAttachmentBody string
-		invalidAttachmentBody string
-		expectedType          blip.MessageType
-		expectedErrorCode     string
-	}{
-		{
-			name:                  "truncated",
-			correctAttachmentBody: "attach",
-			invalidAttachmentBody: "att", // truncate so length and digest are incorrect
-			expectedType:          blip.ErrorType,
-			expectedErrorCode:     strconv.Itoa(http.StatusBadRequest),
-		},
-		{
-			name:                  "malformed",
-			correctAttachmentBody: "attach",
-			invalidAttachmentBody: "attahc", // swap two chars so only digest doesn't match
-			expectedType:          blip.ErrorType,
-			expectedErrorCode:     strconv.Itoa(http.StatusBadRequest),
-		},
-		{
-			name:                  "correct",
-			correctAttachmentBody: "attach",
-			invalidAttachmentBody: "attach",
-			expectedType:          blip.ResponseType,
-			expectedErrorCode:     "",
-		},
-	}
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyHTTP, base.KeySync, base.KeySyncMsg)
-
-	// Create blip tester
-	bt, err := NewBlipTesterFromSpec(t, BlipTesterSpec{
-		connectingUsername:          "user1",
-		connectingPassword:          "1234",
-		connectingUserChannelGrants: []string{"*"}, // All channels
-	})
-	require.NoError(t, err, "Unexpected error creating BlipTester")
-	defer bt.Close()
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			digest := db.Sha1DigestKey([]byte(test.correctAttachmentBody))
-
-			// Send revision with attachment
-			input := SendRevWithAttachmentInput{
-				docId:            test.name,
-				revId:            "1-rev1",
-				attachmentName:   "myAttachment",
-				attachmentLength: len(test.correctAttachmentBody),
-				attachmentBody:   test.invalidAttachmentBody,
-				attachmentDigest: digest,
-			}
-			sent, _, resp := bt.SendRevWithAttachment(input)
-			assert.True(t, sent)
-
-			// Make sure we get the expected response back
-			assert.Equal(t, test.expectedType, resp.Type())
-			if test.expectedErrorCode != "" {
-				assert.Equal(t, test.expectedErrorCode, resp.Properties["Error-Code"])
-			}
-
-			respBody, err := resp.Body()
-			assert.NoError(t, err)
-			t.Logf("resp.Body: %v", string(respBody))
-		})
-	}
 }
 
 // Put a revision that is rejected by the sync function and assert that Sync Gateway
@@ -2744,106 +2558,6 @@ func TestBlipNonDeltaSyncPush(t *testing.T) {
 	assert.Contains(t, resp.Body.String(), `{"howdy":"bob"}`)
 }
 
-// TestBlipDeltaSyncNewAttachmentPull tests that adding a new attachment in SG and replicated via delta sync adds the attachment
-// to the temporary "allowedAttachments" map.
-func TestBlipDeltaSyncNewAttachmentPull(t *testing.T) {
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyAll)
-
-	sgUseDeltas := base.IsEnterpriseEdition()
-	rtConfig := RestTesterConfig{
-		DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
-			DeltaSync: &DeltaSyncConfig{
-				Enabled: &sgUseDeltas,
-			},
-		}},
-		GuestEnabled: true,
-	}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	client, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer client.Close()
-
-	client.ClientDeltas = true
-	err = client.StartPull()
-	assert.NoError(t, err)
-
-	// create doc1 rev 1-0335a345b6ffed05707ccc4cbc1b67f4
-	resp := rt.SendAdminRequest(http.MethodPut, "/db/doc1", `{"greetings": [{"hello": "world!"}, {"hi": "alice"}]}`)
-	assert.Equal(t, http.StatusCreated, resp.Code)
-
-	data, ok := client.WaitForRev("doc1", "1-0335a345b6ffed05707ccc4cbc1b67f4")
-	assert.True(t, ok)
-	assert.Equal(t, `{"greetings":[{"hello":"world!"},{"hi":"alice"}]}`, string(data))
-
-	// create doc1 rev 2-10000d5ec533b29b117e60274b1e3653 on SG with the first attachment
-	resp = rt.SendAdminRequest(http.MethodPut, "/db/doc1?rev=1-0335a345b6ffed05707ccc4cbc1b67f4", `{"greetings": [{"hello": "world!"}, {"hi": "alice"}], "_attachments": {"hello.txt": {"data":"aGVsbG8gd29ybGQ="}}}`)
-	assert.Equal(t, http.StatusCreated, resp.Code)
-	fmt.Println(resp.Body.String())
-
-	data, ok = client.WaitForRev("doc1", "2-10000d5ec533b29b117e60274b1e3653")
-	assert.True(t, ok)
-	var dataMap map[string]interface{}
-	assert.NoError(t, base.JSONUnmarshal(data, &dataMap))
-	atts, ok := dataMap[db.BodyAttachments].(map[string]interface{})
-	require.True(t, ok)
-	assert.Len(t, atts, 1)
-	hello, ok := atts["hello.txt"].(map[string]interface{})
-	require.True(t, ok)
-	assert.Equal(t, "sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=", hello["digest"])
-	assert.Equal(t, float64(11), hello["length"])
-	assert.Equal(t, float64(2), hello["revpos"])
-	assert.Equal(t, true, hello["stub"])
-
-	// message #3 is the getAttachment message that is sent in-between rev processing
-	msg, ok := client.pullReplication.WaitForMessage(3)
-	assert.True(t, ok)
-	assert.NotEqual(t, blip.ErrorType, msg.Type(), "Expected non-error blip message type")
-
-	// Check EE is delta, and CE is full-body replication
-	// msg, ok = client.pullReplication.WaitForMessage(5)
-	msg, ok = client.WaitForBlipRevMessage("doc1", "2-10000d5ec533b29b117e60274b1e3653")
-	assert.True(t, ok)
-
-	if base.IsEnterpriseEdition() {
-		// Check the request was sent with the correct deltaSrc property
-		assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg.Properties[db.RevMessageDeltaSrc])
-		// Check the request body was the actual delta
-		msgBody, err := msg.Body()
-		assert.NoError(t, err)
-		assert.Equal(t, `{"_attachments":[{"hello.txt":{"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=","length":11,"revpos":2,"stub":true}}]}`, string(msgBody))
-	} else {
-		// Check the request was NOT sent with a deltaSrc property
-		assert.Equal(t, "", msg.Properties[db.RevMessageDeltaSrc])
-		// Check the request body was NOT the delta
-		msgBody, err := msg.Body()
-		assert.NoError(t, err)
-		assert.NotEqual(t, `{"_attachments":[{"hello.txt":{"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=","length":11,"revpos":2,"stub":true}}]}`, string(msgBody))
-		assert.Contains(t, string(msgBody), `"_attachments":{"hello.txt":{"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=","length":11,"revpos":2,"stub":true}}`)
-		assert.Contains(t, string(msgBody), `"greetings":[{"hello":"world!"},{"hi":"alice"}]`)
-	}
-
-	resp = rt.SendAdminRequest(http.MethodGet, "/db/doc1?rev=2-10000d5ec533b29b117e60274b1e3653", "")
-	assert.Equal(t, http.StatusOK, resp.Code)
-	var respBody db.Body
-	assert.NoError(t, base.JSONUnmarshal(resp.Body.Bytes(), &respBody))
-	assert.Equal(t, "doc1", respBody[db.BodyId])
-	assert.Equal(t, "2-10000d5ec533b29b117e60274b1e3653", respBody[db.BodyRev])
-	greetings := respBody["greetings"].([]interface{})
-	assert.Len(t, greetings, 2)
-	assert.Equal(t, map[string]interface{}{"hello": "world!"}, greetings[0])
-	assert.Equal(t, map[string]interface{}{"hi": "alice"}, greetings[1])
-	atts = respBody[db.BodyAttachments].(map[string]interface{})
-	assert.Len(t, atts, 1)
-	assert.Equal(t, "sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=", hello["digest"])
-	assert.Equal(t, float64(11), hello["length"])
-	assert.Equal(t, float64(2), hello["revpos"])
-	assert.Equal(t, true, hello["stub"])
-
-	// assert.Equal(t, `{"_attachments":{"hello.txt":{"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=","length":11,"revpos":2,"stub":true}},"_id":"doc1","_rev":"2-10000d5ec533b29b117e60274b1e3653","greetings":[{"hello":"world!"},{"hi":"alice"}]}`, resp.Body.String())
-}
-
 // Reproduces CBG-617 (a client using activeOnly for the initial replication, and then still expecting to get subsequent tombstones afterwards)
 func TestActiveOnlyContinuous(t *testing.T) {
 
@@ -2925,506 +2639,6 @@ func TestNoRevSetSeq(t *testing.T) {
 
 	norevMsg.SetSeq(db.SequenceID{Seq: 60})
 	assert.Equal(t, "60", norevMsg.Properties[db.NorevMessageSeq])
-
-}
-
-// TestBlipDeltaSyncPushAttachment tests updating a doc that has an attachment with a delta that doesn't modify the attachment.
-func TestBlipDeltaSyncPushAttachment(t *testing.T) {
-	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAll)
-
-	if !base.IsEnterpriseEdition() {
-		t.Skip("Delta test requires EE")
-	}
-
-	const docID = "pushAttachmentDoc"
-
-	rt := NewRestTester(t, &RestTesterConfig{
-		DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
-			DeltaSync: &DeltaSyncConfig{
-				Enabled: base.BoolPtr(true),
-			},
-		}},
-		GuestEnabled: true,
-	})
-	defer rt.Close()
-
-	btc, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer btc.Close()
-
-	// Push first rev
-	revID, err := btc.PushRev(docID, "", []byte(`{"key":"val"}`))
-	require.NoError(t, err)
-
-	// Push second rev with an attachment (no delta yet)
-	attData := base64.StdEncoding.EncodeToString([]byte("attach"))
-	revID, err = btc.PushRev(docID, revID, []byte(`{"key":"val","_attachments":{"myAttachment":{"data":"`+attData+`"}}}`))
-	require.NoError(t, err)
-
-	syncData, err := rt.GetDatabase().GetSingleDatabaseCollection().GetDocSyncData(base.TestCtx(t), docID)
-	require.NoError(t, err)
-
-	assert.Len(t, syncData.Attachments, 1)
-	_, found := syncData.Attachments["myAttachment"]
-	assert.True(t, found)
-
-	// Turn deltas on
-	btc.ClientDeltas = true
-
-	// Get existing body with the stub attachment, insert a new property and push as delta.
-	body, found := btc.GetRev(docID, revID)
-	require.True(t, found)
-	newBody, err := base.InjectJSONPropertiesFromBytes(body, base.KVPairBytes{Key: "update", Val: []byte(`true`)})
-	require.NoError(t, err)
-	revID, err = btc.PushRev(docID, revID, newBody)
-	require.NoError(t, err)
-
-	syncData, err = rt.GetDatabase().GetSingleDatabaseCollection().GetDocSyncData(base.TestCtx(t), docID)
-	require.NoError(t, err)
-
-	assert.Len(t, syncData.Attachments, 1)
-	_, found = syncData.Attachments["myAttachment"]
-	assert.True(t, found)
-}
-
-// Test pushing and pulling new attachments through delta sync
-// 1. Create test client that have deltas enabled
-// 2. Start continuous push and pull replication in client
-// 3. Make sure that sync gateway is running with delta sync on, in enterprise edition
-// 4. Create doc with attachment in SGW
-// 5. Update doc in the test client by adding another attachment
-// 6. Have that update pushed using delta sync via the continuous replication started in step 2
-func TestBlipDeltaSyncPushPullNewAttachment(t *testing.T) {
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyAll)
-	if !base.IsEnterpriseEdition() {
-		t.Skip("Delta test requires EE")
-	}
-	rtConfig := RestTesterConfig{
-		DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
-			DeltaSync: &DeltaSyncConfig{
-				Enabled: base.BoolPtr(true),
-			},
-		}},
-		GuestEnabled: true,
-	}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	btc, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer btc.Close()
-
-	btc.ClientDeltas = true
-	err = btc.StartPull()
-	assert.NoError(t, err)
-	const docID = "doc1"
-
-	// Create doc1 rev 1-77d9041e49931ceef58a1eef5fd032e8 on SG with an attachment
-	bodyText := `{"greetings":[{"hi": "alice"}],"_attachments":{"hello.txt":{"data":"aGVsbG8gd29ybGQ="}}}`
-	response := rt.SendAdminRequest(http.MethodPut, "/db/"+docID, bodyText)
-	assert.Equal(t, http.StatusCreated, response.Code)
-
-	// Wait for the document to be replicated at the client
-	revId := RespRevID(t, response)
-	data, ok := btc.WaitForRev(docID, revId)
-	assert.True(t, ok)
-	bodyTextExpected := `{"greetings":[{"hi":"alice"}],"_attachments":{"hello.txt":{"revpos":1,"length":11,"stub":true,"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0="}}}`
-	require.JSONEq(t, bodyTextExpected, string(data))
-
-	// Update the replicated doc at client by adding another attachment.
-	bodyText = `{"greetings":[{"hi":"alice"}],"_attachments":{"hello.txt":{"revpos":1,"length":11,"stub":true,"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0="},"world.txt":{"data":"bGVsbG8gd29ybGQ="}}}`
-	revId, err = btc.PushRev(docID, revId, []byte(bodyText))
-	require.NoError(t, err)
-	assert.Equal(t, "2-abc", revId)
-
-	// Wait for the document to be replicated at SG
-	_, ok = btc.pushReplication.WaitForMessage(2)
-	assert.True(t, ok)
-
-	resp := rt.SendAdminRequest(http.MethodGet, "/db/"+docID+"?rev="+revId, "")
-	assert.Equal(t, http.StatusOK, resp.Code)
-	var respBody db.Body
-	assert.NoError(t, base.JSONUnmarshal(resp.Body.Bytes(), &respBody))
-
-	assert.Equal(t, docID, respBody[db.BodyId])
-	assert.Equal(t, "2-abc", respBody[db.BodyRev])
-	greetings := respBody["greetings"].([]interface{})
-	assert.Len(t, greetings, 1)
-	assert.Equal(t, map[string]interface{}{"hi": "alice"}, greetings[0])
-
-	attachments, ok := respBody[db.BodyAttachments].(map[string]interface{})
-	require.True(t, ok)
-	assert.Len(t, attachments, 2)
-	hello, ok := attachments["hello.txt"].(map[string]interface{})
-	require.True(t, ok)
-	assert.Equal(t, "sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=", hello["digest"])
-	assert.Equal(t, float64(11), hello["length"])
-	assert.Equal(t, float64(1), hello["revpos"])
-	assert.Equal(t, true, hello["stub"])
-
-	world, ok := attachments["world.txt"].(map[string]interface{})
-	require.True(t, ok)
-	assert.Equal(t, "sha1-qiF39gVoGPFzpRQkNYcY9u3wx9Y=", world["digest"])
-	assert.Equal(t, float64(11), world["length"])
-	assert.Equal(t, float64(2), world["revpos"])
-	assert.Equal(t, true, world["stub"])
-}
-
-// Test pushing and pulling v2 attachments with v2 client
-// 1. Create test client.
-// 2. Start continuous push and pull replication in client
-// 3. Create doc with attachment in SGW
-// 4. Update doc in the test client and keep the same attachment stub.
-// 5. Have that update pushed via the continuous replication started in step 2
-func TestBlipPushPullV2AttachmentV2Client(t *testing.T) {
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyAll)
-	rtConfig := RestTesterConfig{
-		DatabaseConfig: &DatabaseConfig{
-			DbConfig: DbConfig{
-				DeltaSync: &DeltaSyncConfig{
-					Enabled: base.BoolPtr(true),
-				},
-			},
-		},
-		GuestEnabled: true,
-	}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	opts := &BlipTesterClientOpts{}
-	opts.SupportedBLIPProtocols = []string{db.BlipCBMobileReplicationV2}
-	btc, err := NewBlipTesterClientOptsWithRT(t, rt, opts)
-	require.NoError(t, err)
-	defer btc.Close()
-
-	err = btc.StartPull()
-	assert.NoError(t, err)
-	const docID = "doc1"
-
-	// Create doc revision with attachment on SG.
-	bodyText := `{"greetings":[{"hi": "alice"}],"_attachments":{"hello.txt":{"data":"aGVsbG8gd29ybGQ="}}}`
-	response := rt.SendAdminRequest(http.MethodPut, "/db/"+docID, bodyText)
-	assert.Equal(t, http.StatusCreated, response.Code)
-
-	// Wait for the document to be replicated to client.
-	revId := RespRevID(t, response)
-	data, ok := btc.WaitForRev(docID, revId)
-	assert.True(t, ok)
-	bodyTextExpected := `{"greetings":[{"hi":"alice"}],"_attachments":{"hello.txt":{"revpos":1,"length":11,"stub":true,"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0="}}}`
-	require.JSONEq(t, bodyTextExpected, string(data))
-
-	// Update the replicated doc at client along with keeping the same attachment stub.
-	bodyText = `{"greetings":[{"hi":"bob"}],"_attachments":{"hello.txt":{"revpos":1,"length":11,"stub":true,"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0="}}}`
-	revId, err = btc.PushRev(docID, revId, []byte(bodyText))
-	require.NoError(t, err)
-	assert.Equal(t, "2-abc", revId)
-
-	// Wait for the document to be replicated at SG
-	_, ok = btc.pushReplication.WaitForMessage(2)
-	assert.True(t, ok)
-
-	resp := rt.SendAdminRequest(http.MethodGet, "/db/"+docID+"?rev="+revId, "")
-	assert.Equal(t, http.StatusOK, resp.Code)
-	var respBody db.Body
-	assert.NoError(t, base.JSONUnmarshal(resp.Body.Bytes(), &respBody))
-
-	assert.Equal(t, docID, respBody[db.BodyId])
-	assert.Equal(t, "2-abc", respBody[db.BodyRev])
-	greetings := respBody["greetings"].([]interface{})
-	assert.Len(t, greetings, 1)
-	assert.Equal(t, map[string]interface{}{"hi": "bob"}, greetings[0])
-
-	attachments, ok := respBody[db.BodyAttachments].(map[string]interface{})
-	require.True(t, ok)
-	assert.Len(t, attachments, 1)
-	hello, ok := attachments["hello.txt"].(map[string]interface{})
-	require.True(t, ok)
-	assert.Equal(t, "sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=", hello["digest"])
-	assert.Equal(t, float64(11), hello["length"])
-	assert.Equal(t, float64(1), hello["revpos"])
-	assert.True(t, hello["stub"].(bool))
-
-	assert.Equal(t, int64(1), rt.GetDatabase().DbStats.CBLReplicationPush().AttachmentPushCount.Value())
-	assert.Equal(t, int64(11), rt.GetDatabase().DbStats.CBLReplicationPush().AttachmentPushBytes.Value())
-}
-
-// Test pushing and pulling v2 attachments with v3 client
-// 1. Create test client.
-// 2. Start continuous push and pull replication in client
-// 3. Create doc with attachment in SGW
-// 4. Update doc in the test client and keep the same attachment stub.
-// 5. Have that update pushed via the continuous replication started in step 2
-func TestBlipPushPullV2AttachmentV3Client(t *testing.T) {
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyAll)
-	rtConfig := RestTesterConfig{
-		DatabaseConfig: &DatabaseConfig{
-			DbConfig: DbConfig{
-				DeltaSync: &DeltaSyncConfig{
-					Enabled: base.BoolPtr(true),
-				},
-			},
-		},
-		GuestEnabled: true,
-	}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	btc, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer btc.Close()
-
-	err = btc.StartPull()
-	assert.NoError(t, err)
-	const docID = "doc1"
-
-	// Create doc revision with attachment on SG.
-	bodyText := `{"greetings":[{"hi": "alice"}],"_attachments":{"hello.txt":{"data":"aGVsbG8gd29ybGQ="}}}`
-	response := rt.SendAdminRequest(http.MethodPut, "/db/"+docID, bodyText)
-	assert.Equal(t, http.StatusCreated, response.Code)
-
-	// Wait for the document to be replicated to client.
-	revId := RespRevID(t, response)
-	data, ok := btc.WaitForRev(docID, revId)
-	assert.True(t, ok)
-	bodyTextExpected := `{"greetings":[{"hi":"alice"}],"_attachments":{"hello.txt":{"revpos":1,"length":11,"stub":true,"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0="}}}`
-	require.JSONEq(t, bodyTextExpected, string(data))
-
-	// Update the replicated doc at client along with keeping the same attachment stub.
-	bodyText = `{"greetings":[{"hi":"bob"}],"_attachments":{"hello.txt":{"revpos":1,"length":11,"stub":true,"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0="}}}`
-	revId, err = btc.PushRev(docID, revId, []byte(bodyText))
-	require.NoError(t, err)
-	assert.Equal(t, "2-abc", revId)
-
-	// Wait for the document to be replicated at SG
-	_, ok = btc.pushReplication.WaitForMessage(2)
-	assert.True(t, ok)
-
-	resp := rt.SendAdminRequest(http.MethodGet, "/db/"+docID+"?rev="+revId, "")
-	assert.Equal(t, http.StatusOK, resp.Code)
-	var respBody db.Body
-	assert.NoError(t, base.JSONUnmarshal(resp.Body.Bytes(), &respBody))
-
-	assert.Equal(t, docID, respBody[db.BodyId])
-	assert.Equal(t, "2-abc", respBody[db.BodyRev])
-	greetings := respBody["greetings"].([]interface{})
-	assert.Len(t, greetings, 1)
-	assert.Equal(t, map[string]interface{}{"hi": "bob"}, greetings[0])
-
-	attachments, ok := respBody[db.BodyAttachments].(map[string]interface{})
-	require.True(t, ok)
-	assert.Len(t, attachments, 1)
-	hello, ok := attachments["hello.txt"].(map[string]interface{})
-	require.True(t, ok)
-	assert.Equal(t, "sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=", hello["digest"])
-	assert.Equal(t, float64(11), hello["length"])
-	assert.Equal(t, float64(1), hello["revpos"])
-	assert.True(t, hello["stub"].(bool))
-
-	assert.Equal(t, int64(1), rt.GetDatabase().DbStats.CBLReplicationPush().AttachmentPushCount.Value())
-	assert.Equal(t, int64(11), rt.GetDatabase().DbStats.CBLReplicationPush().AttachmentPushBytes.Value())
-}
-
-func TestUpdateExistingAttachment(t *testing.T) {
-	rt := NewRestTester(t, &RestTesterConfig{
-		GuestEnabled: true,
-	})
-	defer rt.Close()
-
-	btc, err := NewBlipTesterClient(t, rt)
-	assert.NoError(t, err)
-	defer btc.Close()
-
-	var doc1Body db.Body
-	var doc2Body db.Body
-
-	// Add doc1 and doc2
-	req := rt.SendAdminRequest("PUT", "/db/doc1", `{}`)
-	RequireStatus(t, req, http.StatusCreated)
-	doc1Bytes := req.BodyBytes()
-	req = rt.SendAdminRequest("PUT", "/db/doc2", `{}`)
-	RequireStatus(t, req, http.StatusCreated)
-	doc2Bytes := req.BodyBytes()
-
-	require.NoError(t, rt.WaitForPendingChanges())
-
-	err = json.Unmarshal(doc1Bytes, &doc1Body)
-	assert.NoError(t, err)
-	err = json.Unmarshal(doc2Bytes, &doc2Body)
-	assert.NoError(t, err)
-
-	err = btc.StartOneshotPull()
-	assert.NoError(t, err)
-
-	_, ok := btc.WaitForRev("doc1", "1-ca9ad22802b66f662ff171f226211d5c")
-	require.True(t, ok)
-	_, ok = btc.WaitForRev("doc2", "1-ca9ad22802b66f662ff171f226211d5c")
-	require.True(t, ok)
-
-	attachmentAData := base64.StdEncoding.EncodeToString([]byte("attachmentA"))
-	attachmentBData := base64.StdEncoding.EncodeToString([]byte("attachmentB"))
-
-	revIDDoc1, err := btc.PushRev("doc1", doc1Body["rev"].(string), []byte(`{"key": "val", "_attachments": {"attachment": {"data": "`+attachmentAData+`"}}}`))
-	require.NoError(t, err)
-	revIDDoc2, err := btc.PushRev("doc2", doc2Body["rev"].(string), []byte(`{"key": "val", "_attachments": {"attachment": {"data": "`+attachmentBData+`"}}}`))
-	require.NoError(t, err)
-
-	err = rt.WaitForRev("doc1", revIDDoc1)
-	assert.NoError(t, err)
-	err = rt.WaitForRev("doc2", revIDDoc2)
-	assert.NoError(t, err)
-
-	_, err = rt.GetDatabase().GetSingleDatabaseCollection().GetDocument(base.TestCtx(t), "doc1", db.DocUnmarshalAll)
-	_, err = rt.GetDatabase().GetSingleDatabaseCollection().GetDocument(base.TestCtx(t), "doc2", db.DocUnmarshalAll)
-
-	revIDDoc1, err = btc.PushRev("doc1", revIDDoc1, []byte(`{"key": "val", "_attachments":{"attachment":{"digest":"sha1-SKk0IV40XSHW37d3H0xpv2+z9Ck=","length":11,"content_type":"","stub":true,"revpos":3}}}`))
-	require.NoError(t, err)
-
-	err = rt.WaitForRev("doc1", revIDDoc1)
-	assert.NoError(t, err)
-
-	doc1, err := rt.GetDatabase().GetSingleDatabaseCollection().GetDocument(base.TestCtx(t), "doc1", db.DocUnmarshalAll)
-	assert.NoError(t, err)
-
-	assert.Equal(t, "sha1-SKk0IV40XSHW37d3H0xpv2+z9Ck=", doc1.Attachments["attachment"].(map[string]interface{})["digest"])
-
-	req = rt.SendAdminRequest("GET", "/db/doc1/attachment", "")
-	assert.Equal(t, "attachmentB", string(req.BodyBytes()))
-}
-
-// TestCBLRevposHandling mimics CBL 2.x's revpos handling (setting incoming revpos to the incoming generation).  Test
-// validates that proveAttachment isn't being invoked when the attachment is already present and the
-// digest doesn't change, regardless of revpos.
-func TestCBLRevposHandling(t *testing.T) {
-	rt := NewRestTester(t, &RestTesterConfig{
-		GuestEnabled: true,
-	})
-	defer rt.Close()
-
-	btc, err := NewBlipTesterClient(t, rt)
-	assert.NoError(t, err)
-	defer btc.Close()
-
-	var doc1Body db.Body
-	var doc2Body db.Body
-
-	// Add doc1 and doc2
-	req := rt.SendAdminRequest("PUT", "/db/doc1", `{}`)
-	RequireStatus(t, req, http.StatusCreated)
-	doc1Bytes := req.BodyBytes()
-	req = rt.SendAdminRequest("PUT", "/db/doc2", `{}`)
-	RequireStatus(t, req, http.StatusCreated)
-	doc2Bytes := req.BodyBytes()
-
-	require.NoError(t, rt.WaitForPendingChanges())
-
-	err = json.Unmarshal(doc1Bytes, &doc1Body)
-	assert.NoError(t, err)
-	err = json.Unmarshal(doc2Bytes, &doc2Body)
-	assert.NoError(t, err)
-
-	err = btc.StartOneshotPull()
-	assert.NoError(t, err)
-
-	_, ok := btc.WaitForRev("doc1", "1-ca9ad22802b66f662ff171f226211d5c")
-	require.True(t, ok)
-	_, ok = btc.WaitForRev("doc2", "1-ca9ad22802b66f662ff171f226211d5c")
-	require.True(t, ok)
-
-	attachmentAData := base64.StdEncoding.EncodeToString([]byte("attachmentA"))
-	attachmentBData := base64.StdEncoding.EncodeToString([]byte("attachmentB"))
-
-	revIDDoc1, err := btc.PushRev("doc1", doc1Body["rev"].(string), []byte(`{"key": "val", "_attachments": {"attachment": {"data": "`+attachmentAData+`"}}}`))
-	require.NoError(t, err)
-	revIDDoc2, err := btc.PushRev("doc2", doc2Body["rev"].(string), []byte(`{"key": "val", "_attachments": {"attachment": {"data": "`+attachmentBData+`"}}}`))
-	require.NoError(t, err)
-
-	err = rt.WaitForRev("doc1", revIDDoc1)
-	assert.NoError(t, err)
-	err = rt.WaitForRev("doc2", revIDDoc2)
-	assert.NoError(t, err)
-
-	_, err = rt.GetDatabase().GetSingleDatabaseCollection().GetDocument(base.TestCtx(t), "doc1", db.DocUnmarshalAll)
-	_, err = rt.GetDatabase().GetSingleDatabaseCollection().GetDocument(base.TestCtx(t), "doc2", db.DocUnmarshalAll)
-
-	// Update doc1, don't change attachment, use correct revpos
-	revIDDoc1, err = btc.PushRev("doc1", revIDDoc1, []byte(`{"key": "val", "_attachments":{"attachment":{"digest":"sha1-wzp8ZyykdEuZ9GuqmxQ7XDrY7Co=","length":11,"content_type":"","stub":true,"revpos":2}}}`))
-	require.NoError(t, err)
-
-	err = rt.WaitForRev("doc1", revIDDoc1)
-	assert.NoError(t, err)
-
-	// Update doc1, don't change attachment, use revpos=generation of revid, as CBL 2.x does.  Should not proveAttachment on digest match.
-	revIDDoc1, err = btc.PushRev("doc1", revIDDoc1, []byte(`{"key": "val", "_attachments":{"attachment":{"digest":"sha1-wzp8ZyykdEuZ9GuqmxQ7XDrY7Co=","length":11,"content_type":"","stub":true,"revpos":4}}}`))
-	require.NoError(t, err)
-
-	// Validate attachment exists
-	attResponse := rt.SendAdminRequest("GET", "/db/doc1/attachment", "")
-	assert.Equal(t, 200, attResponse.Code)
-	assert.Equal(t, "attachmentA", string(attResponse.BodyBytes()))
-
-	attachmentPushCount := rt.GetDatabase().DbStats.CBLReplicationPushStats.AttachmentPushCount.Value()
-	// Update doc1, change attachment digest with CBL revpos=generation.  Should getAttachment
-	_, err = btc.PushRev("doc1", revIDDoc1, []byte(`{"key": "val", "_attachments":{"attachment":{"digest":"sha1-SKk0IV40XSHW37d3H0xpv2+z9Ck=","length":11,"content_type":"","stub":true,"revpos":5}}}`))
-	require.NoError(t, err)
-
-	// Validate attachment exists and is updated
-	attResponse = rt.SendAdminRequest("GET", "/db/doc1/attachment", "")
-	assert.Equal(t, 200, attResponse.Code)
-	assert.Equal(t, "attachmentB", string(attResponse.BodyBytes()))
-
-	attachmentPushCountAfter := rt.GetDatabase().DbStats.CBLReplicationPushStats.AttachmentPushCount.Value()
-	assert.Equal(t, attachmentPushCount+1, attachmentPushCountAfter)
-
-}
-
-// TestPushUnknownAttachmentAsStub sets revpos to an older generation, for an attachment that doesn't exist on the server.
-// Verifies that getAttachment is triggered, and attachment is properly persisted.
-func TestPushUnknownAttachmentAsStub(t *testing.T) {
-	rt := NewRestTester(t, &RestTesterConfig{
-		GuestEnabled: true,
-	})
-	defer rt.Close()
-
-	btc, err := NewBlipTesterClient(t, rt)
-	assert.NoError(t, err)
-	defer btc.Close()
-
-	var doc1Body db.Body
-
-	// Add doc1 and doc2
-	req := rt.SendAdminRequest("PUT", "/db/doc1", `{}`)
-	RequireStatus(t, req, http.StatusCreated)
-	doc1Bytes := req.BodyBytes()
-
-	require.NoError(t, rt.WaitForPendingChanges())
-
-	err = json.Unmarshal(doc1Bytes, &doc1Body)
-	assert.NoError(t, err)
-
-	err = btc.StartOneshotPull()
-	assert.NoError(t, err)
-
-	rev1ID := "1-ca9ad22802b66f662ff171f226211d5c"
-	_, ok := btc.WaitForRev("doc1", rev1ID)
-	require.True(t, ok)
-
-	// force attachment into test client's store to validate it's fetched
-	attachmentAData := base64.StdEncoding.EncodeToString([]byte("attachmentA"))
-	contentType := "text/plain"
-	length, digest, err := btc.saveAttachment(contentType, attachmentAData)
-	require.NoError(t, err)
-
-	// Update doc1, include reference to non-existing attachment with recent revpos
-	revIDDoc1, err := btc.PushRev("doc1", rev1ID, []byte(fmt.Sprintf(`{"key": "val", "_attachments":{"attachment":{"digest":"%s","length":%d,"content_type":"%s","stub":true,"revpos":1}}}`, digest, length, contentType)))
-	require.NoError(t, err)
-
-	err = rt.WaitForRev("doc1", revIDDoc1)
-	assert.NoError(t, err)
-
-	// verify that attachment exists on document and was persisted
-	attResponse := rt.SendAdminRequest("GET", "/db/doc1/attachment", "")
-	assert.Equal(t, 200, attResponse.Code)
-	assert.Equal(t, "attachmentA", string(attResponse.BodyBytes()))
 
 }
 
@@ -3622,180 +2836,6 @@ func TestRemovedMessageWithAlternateAccessAndChannelFilteredReplication(t *testi
 	require.Equal(t, "docmarker", messageBody[0].([]interface{})[1])
 }
 
-func TestBlipPushPullNewAttachmentCommonAncestor(t *testing.T) {
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyAll)
-	rtConfig := RestTesterConfig{
-		GuestEnabled: true,
-	}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	btc, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer btc.Close()
-
-	err = btc.StartPull()
-	assert.NoError(t, err)
-	const docID = "doc1"
-
-	// CBL creates revisions 1-abc,2-abc on the client, with an attachment associated with rev 2.
-	bodyText := `{"greetings":[{"hi":"alice"}],"_attachments":{"hello.txt":{"data":"aGVsbG8gd29ybGQ="}}}`
-	err = btc.StoreRevOnClient(docID, "2-abc", []byte(bodyText))
-	require.NoError(t, err)
-
-	bodyText = `{"greetings":[{"hi":"alice"}],"_attachments":{"hello.txt":{"revpos":2,"length":11,"stub":true,"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0="}}}`
-	revId, err := btc.PushRevWithHistory(docID, "", []byte(bodyText), 2, 0)
-	require.NoError(t, err)
-	assert.Equal(t, "2-abc", revId)
-
-	// Wait for the documents to be replicated at SG
-	_, ok := btc.pushReplication.WaitForMessage(2)
-	assert.True(t, ok)
-
-	resp := rt.SendAdminRequest(http.MethodGet, "/db/"+docID+"?rev="+revId, "")
-	assert.Equal(t, http.StatusOK, resp.Code)
-
-	// CBL updates the doc w/ two more revisions, 3-abc, 4-abc,
-	// these are sent to SG as 4-abc, history:[4-abc,3-abc,2-abc], the attachment has revpos=2
-	bodyText = `{"greetings":[{"hi":"bob"}],"_attachments":{"hello.txt":{"revpos":2,"length":11,"stub":true,"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0="}}}`
-	revId, err = btc.PushRevWithHistory(docID, revId, []byte(bodyText), 2, 0)
-	require.NoError(t, err)
-	assert.Equal(t, "4-abc", revId)
-
-	// Wait for the document to be replicated at SG
-	_, ok = btc.pushReplication.WaitForMessage(4)
-	assert.True(t, ok)
-
-	resp = rt.SendAdminRequest(http.MethodGet, "/db/"+docID+"?rev="+revId, "")
-	assert.Equal(t, http.StatusOK, resp.Code)
-
-	var respBody db.Body
-	assert.NoError(t, base.JSONUnmarshal(resp.Body.Bytes(), &respBody))
-
-	assert.Equal(t, docID, respBody[db.BodyId])
-	assert.Equal(t, "4-abc", respBody[db.BodyRev])
-	greetings := respBody["greetings"].([]interface{})
-	assert.Len(t, greetings, 1)
-	assert.Equal(t, map[string]interface{}{"hi": "bob"}, greetings[0])
-
-	attachments, ok := respBody[db.BodyAttachments].(map[string]interface{})
-	require.True(t, ok)
-	assert.Len(t, attachments, 1)
-	hello, ok := attachments["hello.txt"].(map[string]interface{})
-	require.True(t, ok)
-	assert.Equal(t, "sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=", hello["digest"])
-	assert.Equal(t, float64(11), hello["length"])
-	assert.Equal(t, float64(2), hello["revpos"])
-	assert.True(t, hello["stub"].(bool))
-
-	// Check the number of sendProveAttachment/sendGetAttachment calls.
-	require.NotNil(t, btc.pushReplication.replicationStats)
-	assert.Equal(t, int64(1), btc.pushReplication.replicationStats.GetAttachment.Value())
-	assert.Equal(t, int64(0), btc.pushReplication.replicationStats.ProveAttachment.Value())
-}
-
-func TestBlipPushPullNewAttachmentNoCommonAncestor(t *testing.T) {
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyAll)
-	rtConfig := RestTesterConfig{
-		GuestEnabled: true,
-	}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	btc, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer btc.Close()
-
-	err = btc.StartPull()
-	assert.NoError(t, err)
-	const docID = "doc1"
-
-	// CBL creates revisions 1-abc, 2-abc, 3-abc, 4-abc on the client, with an attachment associated with rev 2.
-	// rev tree pruning on the CBL side, so 1-abc no longer exists.
-	// CBL replicates, sends to client as 4-abc history:[4-abc, 3-abc, 2-abc], attachment has revpos=2
-	bodyText := `{"greetings":[{"hi":"alice"}],"_attachments":{"hello.txt":{"data":"aGVsbG8gd29ybGQ="}}}`
-	err = btc.StoreRevOnClient(docID, "2-abc", []byte(bodyText))
-	require.NoError(t, err)
-
-	bodyText = `{"greetings":[{"hi":"alice"}],"_attachments":{"hello.txt":{"revpos":2,"length":11,"stub":true,"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0="}}}`
-	revId, err := btc.PushRevWithHistory(docID, "2-abc", []byte(bodyText), 2, 0)
-	require.NoError(t, err)
-	assert.Equal(t, "4-abc", revId)
-
-	// Wait for the document to be replicated at SG
-	_, ok := btc.pushReplication.WaitForMessage(2)
-	assert.True(t, ok)
-
-	resp := rt.SendAdminRequest(http.MethodGet, "/db/"+docID+"?rev="+revId, "")
-	assert.Equal(t, http.StatusOK, resp.Code)
-
-	var respBody db.Body
-	assert.NoError(t, base.JSONUnmarshal(resp.Body.Bytes(), &respBody))
-
-	assert.Equal(t, docID, respBody[db.BodyId])
-	assert.Equal(t, "4-abc", respBody[db.BodyRev])
-	greetings := respBody["greetings"].([]interface{})
-	assert.Len(t, greetings, 1)
-	assert.Equal(t, map[string]interface{}{"hi": "alice"}, greetings[0])
-
-	attachments, ok := respBody[db.BodyAttachments].(map[string]interface{})
-	require.True(t, ok)
-	assert.Len(t, attachments, 1)
-	hello, ok := attachments["hello.txt"].(map[string]interface{})
-	require.True(t, ok)
-	assert.Equal(t, "sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=", hello["digest"])
-	assert.Equal(t, float64(11), hello["length"])
-	assert.Equal(t, float64(4), hello["revpos"])
-	assert.True(t, hello["stub"].(bool))
-
-	// Check the number of sendProveAttachment/sendGetAttachment calls.
-	require.NotNil(t, btc.pushReplication.replicationStats)
-	assert.Equal(t, int64(1), btc.pushReplication.replicationStats.GetAttachment.Value())
-	assert.Equal(t, int64(0), btc.pushReplication.replicationStats.ProveAttachment.Value())
-}
-
-func TestMinRevPosWorkToAvoidUnnecessaryProveAttachment(t *testing.T) {
-	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAll)
-	rt := NewRestTester(t, &RestTesterConfig{
-		GuestEnabled: true,
-		DatabaseConfig: &DatabaseConfig{
-			DbConfig: DbConfig{
-				AllowConflicts: base.BoolPtr(true),
-			},
-		},
-	})
-	defer rt.Close()
-
-	btc, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer btc.Close()
-
-	// Push an initial rev with attachment data
-	initialRevID := rt.CreateDocReturnRev(t, "doc", "", map[string]interface{}{"_attachments": map[string]interface{}{"hello.txt": map[string]interface{}{"data": "aGVsbG8gd29ybGQ="}}})
-	err = rt.WaitForPendingChanges()
-	assert.NoError(t, err)
-
-	// Replicate data to client and ensure doc arrives
-	err = btc.StartOneshotPull()
-	assert.NoError(t, err)
-	_, found := btc.WaitForRev("doc", initialRevID)
-	assert.True(t, found)
-
-	// Push a revision with a bunch of history simulating doc updated on mobile device
-	// Note this references revpos 1 and therefore SGW has it - Shouldn't need proveAttachment
-	proveAttachmentBefore := btc.pushReplication.replicationStats.ProveAttachment.Value()
-	revid, err := btc.PushRevWithHistory("doc", initialRevID, []byte(`{"_attachments": {"hello.txt": {"revpos":1,"stub":true,"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0="}}}`), 25, 5)
-	assert.NoError(t, err)
-	proveAttachmentAfter := btc.pushReplication.replicationStats.ProveAttachment.Value()
-	assert.Equal(t, proveAttachmentBefore, proveAttachmentAfter)
-
-	// Push another bunch of history
-	_, err = btc.PushRevWithHistory("doc", revid, []byte(`{"_attachments": {"hello.txt": {"revpos":1,"stub":true,"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0="}}}`), 25, 5)
-	assert.NoError(t, err)
-	proveAttachmentAfter = btc.pushReplication.replicationStats.ProveAttachment.Value()
-	assert.Equal(t, proveAttachmentBefore, proveAttachmentAfter)
-}
-
 // Make sure that a client cannot open multiple subChanges subscriptions on a single blip context (SG #3222)
 // - Open a one-off subChanges request, ensure it works.
 // - Open a subsequent continuous request, and ensure it works.
@@ -3918,46 +2958,6 @@ func TestMultipleOutstandingChangesSubscriptions(t *testing.T) {
 
 	base.RequireWaitForStat(t, pullStats.NumPullReplActiveOneShot.Value, 0)
 	base.RequireWaitForStat(t, pullStats.NumPullReplActiveContinuous.Value, 0)
-}
-
-func TestAttachmentWithErroneousRevPos(t *testing.T) {
-	rt := NewRestTester(t, &RestTesterConfig{
-		GuestEnabled: true,
-	})
-	defer rt.Close()
-
-	btc, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer btc.Close()
-
-	// Create rev 1 with the hello.txt attachment
-	revid := rt.CreateDocReturnRev(t, "doc", "", map[string]interface{}{"val": "val", "_attachments": map[string]interface{}{"hello.txt": map[string]interface{}{"data": "aGVsbG8gd29ybGQ="}}})
-	err = rt.WaitForPendingChanges()
-	assert.NoError(t, err)
-
-	// Pull rev and attachment down to client
-	err = btc.StartOneshotPull()
-	assert.NoError(t, err)
-	_, found := btc.WaitForRev("doc", revid)
-	assert.True(t, found)
-
-	// Add an attachment to client
-	btc.AttachmentsLock().Lock()
-	btc.Attachments()["sha1-l+N7VpXGnoxMm8xfvtWPbz2YvDc="] = []byte("goodbye cruel world")
-	btc.AttachmentsLock().Unlock()
-
-	// Put doc with an erroneous revpos 1 but with a different digest, referring to the above attachment
-	_, err = btc.PushRevWithHistory("doc", revid, []byte(`{"_attachments": {"hello.txt": {"revpos":1,"stub":true,"length": 19,"digest":"sha1-l+N7VpXGnoxMm8xfvtWPbz2YvDc="}}}`), 1, 0)
-	require.NoError(t, err)
-
-	// Ensure message and attachment is pushed up
-	_, ok := btc.pushReplication.WaitForMessage(2)
-	assert.True(t, ok)
-
-	// Get the attachment and ensure the data is updated
-	resp := rt.SendAdminRequest(http.MethodGet, "/db/doc/hello.txt", "")
-	RequireStatus(t, resp, http.StatusOK)
-	assert.Equal(t, "goodbye cruel world", string(resp.BodyBytes()))
 }
 
 func TestBlipInternalPropertiesHandling(t *testing.T) {
@@ -4099,60 +3099,6 @@ func TestBlipInternalPropertiesHandling(t *testing.T) {
 	}
 }
 
-// CBG-2004: Test that prove attachment over Blip works correctly when receiving a ErrAttachmentNotFound
-func TestProveAttachmentNotFound(t *testing.T) {
-	rt := NewRestTester(t, &RestTesterConfig{
-		GuestEnabled: true,
-	})
-	defer rt.Close()
-
-	bt, err := NewBlipTesterFromSpecWithRT(t, nil, rt)
-	assert.NoError(t, err, "Error creating BlipTester")
-	defer bt.Close()
-
-	attachmentData := []byte("attachmentA")
-	attachmentDataEncoded := base64.StdEncoding.EncodeToString(attachmentData)
-
-	bt.blipContext.HandlerForProfile[db.MessageProveAttachment] = func(msg *blip.Message) {
-		status, errMsg := base.ErrorAsHTTPStatus(db.ErrAttachmentNotFound)
-		msg.Response().SetError("HTTP", status, errMsg)
-	}
-
-	// Handler for when full attachment is requested
-	bt.blipContext.HandlerForProfile[db.MessageGetAttachment] = func(msg *blip.Message) {
-		resp := msg.Response()
-		resp.SetBody(attachmentData)
-		resp.SetCompressed(msg.Properties[db.BlipCompress] == "true")
-	}
-
-	// Initial set up
-	sent, _, _, err := bt.SendRev("doc1", "1-abc", []byte(`{"key": "val", "_attachments": {"attachment": {"data": "`+attachmentDataEncoded+`"}}}`), blip.Properties{})
-	require.True(t, sent)
-	require.NoError(t, err)
-
-	err = rt.WaitForPendingChanges()
-	require.NoError(t, err)
-
-	// Should log:
-	// "Peer sent prove attachment error 404 attachment not found, falling back to getAttachment for proof in doc <ud>doc1</ud> (digest sha1-wzp8ZyykdEuZ9GuqmxQ7XDrY7Co=)"
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyAll)
-
-	// Use different attachment name to bypass digest check in ForEachStubAttachment() which skips prove attachment code
-	// Set attachment to V2 so it can be retrieved by RT successfully
-	sent, _, _, err = bt.SendRev("doc1", "2-abc", []byte(`{"key": "val", "_attachments":{"attach":{"digest":"sha1-wzp8ZyykdEuZ9GuqmxQ7XDrY7Co=","length":11,"stub":true,"revpos":1,"ver":2}}}`), blip.Properties{})
-	require.True(t, sent)
-	require.NoError(t, err)
-
-	err = rt.WaitForPendingChanges()
-	require.NoError(t, err)
-	// Check attachment is on the document
-	body := rt.GetDoc("doc1")
-	assert.Equal(t, "2-abc", body.ExtractRev())
-	resp := rt.SendAdminRequest("GET", "/db/doc1/attach", "")
-	RequireStatus(t, resp, 200)
-	assert.EqualValues(t, attachmentData, resp.BodyBytes())
-}
-
 // CBG-2053: Test that the handleRev stats still increment correctly when going through the processRev function with
 // the stat mapping (processRevStats)
 func TestProcessRevIncrementsStat(t *testing.T) {
@@ -4254,170 +3200,6 @@ func TestSendRevAsReadOnlyGuest(t *testing.T) {
 	body, err = revResponse.Body()
 	log.Printf("response body: %s", body)
 
-}
-
-// TestBlipAttachNameChange tests CBL handling - attachments with changed names are sent as stubs, and not new attachments
-func TestBlipAttachNameChange(t *testing.T) {
-	rt := NewRestTester(t, &RestTesterConfig{
-		GuestEnabled: true,
-	})
-	defer rt.Close()
-
-	client1, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer client1.Close()
-	base.SetUpTestLogging(t, base.LevelTrace, base.KeySync, base.KeySyncMsg, base.KeyWebSocket, base.KeyWebSocketFrame, base.KeyHTTP, base.KeyCRUD)
-
-	attachmentA := []byte("attachmentA")
-	attachmentAData := base64.StdEncoding.EncodeToString(attachmentA)
-	digest := db.Sha1DigestKey(attachmentA)
-
-	// Push initial attachment data
-	rev, err := client1.PushRev("doc", "", []byte(`{"key":"val","_attachments":{"attachment": {"data":"`+attachmentAData+`"}}}`))
-	require.NoError(t, err)
-
-	// Confirm attachment is in the bucket
-	attachmentAKey := db.MakeAttachmentKey(2, "doc", digest)
-	bucketAttachmentA, _, err := rt.Bucket().GetRaw(attachmentAKey)
-	require.NoError(t, err)
-	require.EqualValues(t, bucketAttachmentA, attachmentA)
-
-	// Simulate changing only the attachment name over CBL
-	// Use revpos 2 to simulate revpos bug in CBL 2.8 - 3.0.0
-	rev, err = client1.PushRev("doc", rev, []byte(`{"key":"val","_attachments":{"attach":{"revpos":2,"content_type":"","length":11,"stub":true,"digest":"`+digest+`"}}}`))
-	require.NoError(t, err)
-	err = rt.WaitForRev("doc", rev)
-	require.NoError(t, err)
-
-	// Check if attachment is still in bucket
-	bucketAttachmentA, _, err = rt.Bucket().GetRaw(attachmentAKey)
-	assert.NoError(t, err)
-	assert.Equal(t, bucketAttachmentA, attachmentA)
-
-	resp := rt.SendAdminRequest("GET", "/db/doc/attach", "")
-	RequireStatus(t, resp, http.StatusOK)
-	assert.Equal(t, attachmentA, resp.BodyBytes())
-}
-
-// TestBlipLegacyAttachNameChange ensures that CBL name changes for legacy attachments are handled correctly
-func TestBlipLegacyAttachNameChange(t *testing.T) {
-	rt := NewRestTester(t, &RestTesterConfig{
-		GuestEnabled: true,
-	})
-	defer rt.Close()
-
-	client1, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer client1.Close()
-	base.SetUpTestLogging(t, base.LevelTrace, base.KeySync, base.KeySyncMsg, base.KeyWebSocket, base.KeyWebSocketFrame, base.KeyHTTP, base.KeyCRUD)
-
-	// Create document in the bucket with a legacy attachment
-	docID := "doc"
-	attBody := []byte(`hi`)
-	digest := db.Sha1DigestKey(attBody)
-	attKey := db.MakeAttachmentKey(db.AttVersion1, docID, digest)
-	rawDoc := rawDocWithAttachmentAndSyncMeta()
-
-	// Create a document with legacy attachment.
-	CreateDocWithLegacyAttachment(t, rt, docID, rawDoc, attKey, attBody)
-
-	// Get the document and grab the revID.
-	responseBody := rt.GetDoc(docID)
-	revID := responseBody["_rev"].(string)
-	require.NotEmpty(t, revID)
-
-	// Store the document and attachment on the test client
-	err = client1.StoreRevOnClient(docID, revID, rawDoc)
-	require.NoError(t, err)
-	client1.AttachmentsLock().Lock()
-	client1.Attachments()[digest] = attBody
-	client1.AttachmentsLock().Unlock()
-
-	// Confirm attachment is in the bucket
-	attachmentAKey := db.MakeAttachmentKey(1, "doc", digest)
-	bucketAttachmentA, _, err := rt.Bucket().GetRaw(attachmentAKey)
-	require.NoError(t, err)
-	require.EqualValues(t, bucketAttachmentA, attBody)
-
-	// Simulate changing only the attachment name over CBL
-	// Use revpos 2 to simulate revpos bug in CBL 2.8 - 3.0.0
-	revID, err = client1.PushRev("doc", revID, []byte(`{"key":"val","_attachments":{"attach":{"revpos":2,"content_type":"test/plain","length":2,"stub":true,"digest":"`+digest+`"}}}`))
-	require.NoError(t, err)
-	err = rt.WaitForRev("doc", revID)
-	require.NoError(t, err)
-
-	resp := rt.SendAdminRequest("GET", "/db/doc/attach", "")
-	RequireStatus(t, resp, http.StatusOK)
-	assert.Equal(t, attBody, resp.BodyBytes())
-}
-
-// TestBlipLegacyAttachNameChange ensures that CBL updates for documents associated with legacy attachments are handled correctly
-func TestBlipLegacyAttachDocUpdate(t *testing.T) {
-	rt := NewRestTester(t, &RestTesterConfig{
-		GuestEnabled: true,
-	})
-	defer rt.Close()
-
-	client1, err := NewBlipTesterClientOptsWithRT(t, rt, nil)
-	require.NoError(t, err)
-	defer client1.Close()
-	base.SetUpTestLogging(t, base.LevelTrace, base.KeySync, base.KeySyncMsg, base.KeyWebSocket, base.KeyWebSocketFrame, base.KeyHTTP, base.KeyCRUD)
-
-	// Create document in the bucket with a legacy attachment.  Properties here align with rawDocWithAttachmentAndSyncMeta
-	docID := "doc"
-	attBody := []byte(`hi`)
-	digest := db.Sha1DigestKey(attBody)
-	attKey := db.MakeAttachmentKey(db.AttVersion1, docID, digest)
-	attName := "hi.txt"
-	rawDoc := rawDocWithAttachmentAndSyncMeta()
-
-	// Create a document with legacy attachment.
-	CreateDocWithLegacyAttachment(t, rt, docID, rawDoc, attKey, attBody)
-
-	// Get the document and grab the revID.
-	responseBody := rt.GetDoc(docID)
-	revID := responseBody["_rev"].(string)
-	require.NotEmpty(t, revID)
-
-	// Store the document and attachment on the test client
-	err = client1.StoreRevOnClient(docID, revID, rawDoc)
-	require.NoError(t, err)
-	client1.AttachmentsLock().Lock()
-	client1.Attachments()[digest] = attBody
-	client1.AttachmentsLock().Unlock()
-
-	// Confirm attachment is in the bucket
-	attachmentAKey := db.MakeAttachmentKey(1, "doc", digest)
-	bucketAttachmentA, _, err := rt.Bucket().GetRaw(attachmentAKey)
-	require.NoError(t, err)
-	require.EqualValues(t, bucketAttachmentA, attBody)
-
-	// Update the document, leaving body intact
-	revID, err = client1.PushRev("doc", revID, []byte(`{"key":"val1","_attachments":{"`+attName+`":{"revpos":2,"content_type":"text/plain","length":2,"stub":true,"digest":"`+digest+`"}}}`))
-	require.NoError(t, err)
-	err = rt.WaitForRev("doc", revID)
-	require.NoError(t, err)
-
-	resp := rt.SendAdminRequest("GET", "/db/doc/"+attName, "")
-	RequireStatus(t, resp, http.StatusOK)
-	assert.Equal(t, attBody, resp.BodyBytes())
-
-	// Validate that the attachment hasn't been migrated to V2
-	v1Key := db.MakeAttachmentKey(1, "doc", digest)
-	v1Body, _, err := rt.Bucket().GetRaw(v1Key)
-	require.NoError(t, err)
-	require.EqualValues(t, attBody, v1Body)
-
-	v2Key := db.MakeAttachmentKey(2, "doc", digest)
-	_, _, err = rt.Bucket().GetRaw(v2Key)
-	require.Error(t, err)
-	// Confirm correct type of error for both integration test and Walrus
-	if !errors.Is(err, sgbucket.MissingError{Key: v2Key}) {
-		var keyValueErr *gocb.KeyValueError
-		require.True(t, errors.As(err, &keyValueErr))
-		require.Equal(t, keyValueErr.StatusCode, memd.StatusKeyNotFound)
-		require.Equal(t, keyValueErr.DocumentID, v2Key)
-	}
 }
 
 // Tests changes made in CBG-2151 to return errors from sendRevision unless it's a document not found error,
